extract redrawGraph out
